# routes/api.php
<?php

use App\Http\Controllers\Auth\AuthController;
use App\Http\Controllers\Products\ProductsController;
use App\Http\Controllers\Destinations\DestinationsController;
use App\Http\Controllers\Trips\TripsController;
use App\Http\Controllers\Users\UsersController;
use App\Http\Controllers\Accommodations\AccommodationsController;
use App\Http\Controllers\Notifications\NotificationsController;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Route;

/*
|--------------------------------------------------------------------------
| API Routes
|--------------------------------------------------------------------------
|
| Here is where you can register API routes for your application. These
| routes are loaded by the RouteServiceProvider within a group which
| is assigned the "api" middleware group. Enjoy building your API!
|
*/


Route::get('/check-auth', [AuthController::class, 'check_auth']);

//Route::get('/check-auth', function (Request $request) {
//    if ($request->user()) {
//        return response()->json(['user' => $request->user()]);
//    } else {
//        return response()->json(['message' => 'User not logged in', 'user' => $request]);
//    }
//});

Route::group([
    'middleware' => 'api'
], function ($router) {

//    Route::middleware('auth')->get('/check-auth', function (Request $request) {
//        return response()->json(['user' => $request->user()]);
//    });

    /**
     * Authentication Module
     */
    Route::group(['prefix' => 'auth'], function () {
        Route::post('register', [AuthController::class, 'register']);
        Route::post('login', [AuthController::class, 'login']);
        Route::post('logout', [AuthController::class, 'logout']);
        Route::post('refresh', [AuthController::class, 'refresh']);
        Route::get('me', [AuthController::class, 'me']);
    });
//    Route::get('me', [AuthController::class, 'me']);


    /**
     * Products Module
     */
    Route::resource('products', ProductsController::class);
    Route::get('products/view/all', [ProductsController::class, 'indexAll']);
    Route::get('products/view/search', [ProductsController::class, 'search']);

    /**
     * Destinations Module
     */
    Route::resource('destinations', DestinationsController::class);

    /**
     * Users Module
     */
    Route::resource('users', UsersController::class);

    /**
     * Trips Module
     */
    Route::resource('trips', TripsController::class);

    /**
     * Accommodations Module
     */
    Route::resource('accommodations', AccommodationsController::class);

    /**
     * Notifications Module
     */
    Route::resource('notifications', NotificationsController::class);

});


/**
 * Start - Public APIs
 */

/**
 * Destinations Module
 */
Route::get('destinations/view/all', [DestinationsController::class, 'indexAll']);

/**
 * End - Public APIs
 */